script: functions menu 1 and menu 2 Fix and create menu 1 and 2

+++ project/script.py
@@ -2,7 +2,9 @@ import json
 import os
 from PyInquirer import prompt
 
-def list_applications(file_name):
+available_avtives_type = ["Criptocurrencies", "Treasure Securities", "Investment Funds", "Stocks", "DI"]
+
+def load_actives(file_name):
     try: 
         with open(file_name, "r", encoding="utf-8") as infos:
             result = json.load(infos)
@@ -10,20 +12,21 @@ def list_applications(file_name):
         result = {}
     return result
 
-def select_active_type(all_actives):
+def list_select_active_type(all_actives):
     if not all_actives:
+        print("There is no older information, please insert a active type to begin.")
         question = [
             {
                 "type": "list",
                 "name": "active_type",
                 "message": "Which active do you want to check?",
-                "choices": ["Add new one","Quit"],
+                "choices": ["Add new active type","Quit"],
             }
         ]
         answer = prompt(question)
         return answer["active_type"]
     else:
-        actives_list = list(all_actives.keys()) + ["Add new one", "Delet existent one", "Quit"]
+        actives_list = list(all_actives.keys()) + ["Add new active type", "Delete existent active type", "Quit"]
         question = [
             {
                 "type": "list",
@@ -36,41 +39,61 @@ def select_active_type(all_actives):
 
         return answer["active_type"]
 
-def add_active_type(all_actives):
-    question = [
-        {
-            "type": "input",
-            "name": "active_type",
-            "message": "Input new active:",
-        }
-    ]
-    answer = prompt(question)
-    all_actives[answer["active_type"]] = []
-    with open("actives.json", "w", encoding="utf-8") as infos:
-        json.dump(all_actives, infos)
-    print("Active type added successfuly!")
+def add_delete_active_type(all_actives, active_type):
+    if active_type == "Add new active type":
+        question = [
+            {
+                "type": "input",
+                "name": "active_type",
+                "message": "Input new active (" + ", ".join(available_avtives_type) + "):",
+            }
+        ]
+        answer = prompt(question)
+        all_actives[answer["active_type"]] = []
+        with open("actives.json", "w", encoding="utf-8") as infos:
+            json.dump(all_actives, infos)
+        print("Active type added successfuly!")
+    else:
+        question = [
+            {
+                "type": "list",
+                "name": "active_type",
+                "message": "Select which active type do tou want to delete",
+                "choices": list(all_actives.keys()),
+            }
+        ]
+        answer = prompt(question)
+        del all_actives[answer["active_type"]]
+        with open("actives.json", "w", encoding="utf-8") as infos:
+            json.dump(all_actives, infos)
+        print("Active type deleted successfuly!")
     input("Press any key to continue...")
 
-def delet_active_type(all_actives):
+def select_active(all_actives, active_type):
+    options = all_actives[active_type] + ["Add new active", "Delet existent active", "Go back"]
     question = [
         {
             "type": "list",
-            "name": "active_type",
-            "message": "Select which active type do tou want to delete",
-            "choices": list(all_actives.keys()),
+            "name": "active",
+            "message": "Which one from {} do tou want to check?".format(active_type),
+            "choices": options,
         }
     ]
     answer = prompt(question)
-    del all_actives[answer["active_type"]]
-    with open("actives.json", "w", encoding="utf-8") as infos:
-        json.dump(all_actives, infos)
-    print("Active type deleted successfuly!")
-    input("Press any key to continue...")
+
+    return answer["active"]
 
 def main():
     os.system("clear")
-    all_actives = list_applications("actives.json")
-    active_type = select_active_type(all_actives)
+    all_actives = load_actives("actives.json")
+    active_type = list_select_active_type(all_actives)
+    if active_type == "Add new active type" or active_type == "Delete existent active type":
+        add_delete_active_type(all_actives,active_type)
+        main()
+    elif active_type == "Quit":
+        return
+    else:
+        active = select_active(all_actives, active_type)
 
 if __name__ == "__main__":
     main()
